Rename build_timing_processor to job_processor
